backend/requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
google-generativeai
langgraph
langchain-google-genai
python-dotenv
pydantic
pydantic-settings
httpx
